backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
chromadb==0.4.22
langchain==0.1.9
langchain-openai==0.0.5
numpy==1.26.4
tiktoken==0.6.0
gunicorn==21.2.0
